Move domain::Tags::InitialExtents to DomainCreator library
