+++ src/Domain/Tags.cpp
@@ -11,12 +11,6 @@
 namespace domain {
 namespace Tags {
 template <size_t Dim>
-std::vector<std::array<size_t, Dim>> InitialExtents<Dim>::create_from_options(
-    const std::unique_ptr<::DomainCreator<Dim>>& domain_creator) {
-  return domain_creator->initial_extents();
-}
-
-template <size_t Dim>
 std::vector<std::array<size_t, Dim>>
 InitialRefinementLevels<Dim>::create_from_options(
     const std::unique_ptr<::DomainCreator<Dim>>& domain_creator) {
@@ -26,9 +20,6 @@ InitialRefinementLevels<Dim>::create_from_options(
 #define DIM(data) BOOST_PP_TUPLE_ELEM(0, data)
 
 #define INSTANTIATE(_, data)                                              \
-  template std::vector<std::array<size_t, DIM(data)>>                     \
-  InitialExtents<DIM(data)>::create_from_options(                         \
-      const std::unique_ptr<::DomainCreator<DIM(data)>>& domain_creator); \
   template std::vector<std::array<size_t, DIM(data)>>                     \
   InitialRefinementLevels<DIM(data)>::create_from_options(                \
       const std::unique_ptr<::DomainCreator<DIM(data)>>& domain_creator);
