+++ tests/Unit/Elliptic/DiscontinuousGalerkin/Test_DgOperator.cpp
@@ -17,6 +17,7 @@
 #include "Domain/Creators/Rectangle.hpp"
 #include "Domain/Creators/RegisterDerivedWithCharm.hpp"
 #include "Domain/Creators/Tags/Domain.hpp"
+#include "Domain/Creators/Tags/InitialExtents.hpp"
 #include "Domain/FaceNormal.hpp"
 #include "Domain/Structure/InitialElementIds.hpp"
 #include "Domain/Tags/ExternalBoundaryConditions.hpp"
